requirements.txt
flask
flask-cors
yfinance
pandas
numpy
python-dotenv
